set correct project settings for release target fixed conversion warnings added binary for msvc_x64

+++ src/Banking4Summary/Banking4Summary.h
@@ -9,7 +9,7 @@ static int getIndexOfColumn(const std::vector<std::string> &headers, string keyw
 
 	if (itBuchungsdatum != headers.end())
 	{
-		return (itBuchungsdatum - headers.begin());
+		return (static_cast<int>(itBuchungsdatum - headers.begin()));
 	}
 	else
 	{
